Refactor: Refactored pipeline logic out of main.py into extract.py.
